--- app/(login)/dashboard/layout.tsx
@@ -1,28 +1,0 @@
-export const metadata = {
-  title: "Next.js",
-  description: "Generated by Next.js",
-};
-import Header from "@/app/(main)/components/Header/Header";
-import { Ubuntu } from "next/font/google";
-
-const ubuntu = Ubuntu({
-  subsets: ["latin"],
-  weight: ["400", "500", "700"],
-});
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en">
-      <body>
-        <div className={ubuntu.className}>
-          <Header />
-        </div>
-        {children}
-      </body>
-    </html>
-  );
-}
