+++ app/dashboard/layout.tsx
@@ -1,0 +1,32 @@
+export const metadata = {
+  title: "Next.js",
+  description: "Generated by Next.js",
+};
+import Header from "@/app/(main)/components/Header/Header";
+import { Ubuntu, Inter } from "next/font/google";
+import "@/app/globals.css";
+
+const ubuntu = Ubuntu({
+  subsets: ["latin"],
+  weight: ["400", "500", "700"],
+});
+
+const inter = Inter({
+  subsets: ["latin"],
+  weight: ["400", "500", "700"],
+});
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="dashboard-layout">
+      <div className={ubuntu.className}>
+        <Header />
+      </div>
+      <section className={inter.className}>{children}</section>
+    </div>
+  );
+}
